Little refactoring and cleaning
