+++ Volontulo/app/src/main/AndroidManifest.xml
@@ -21,12 +21,14 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".LoginActivity" />
+        <activity
+            android:name=".ui.login.LoginActivity"
+            android:windowSoftInputMode="adjustResize" />
         <activity android:name=".ui.offers.OfferActivity" />
         <activity
             android:name=".ui.offers.AddOfferActivity"
-            android:theme="@style/AppTheme.BorderlessButtons"
-            android:parentActivityName=".ui.offers.OfferActivity">
+            android:parentActivityName=".ui.offers.OfferActivity"
+            android:theme="@style/AppTheme.BorderlessButtons">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".ui.offers.OfferActivity" />
