+++ Volontulo/app/src/main/java/com/stxnext/volontulo/VolontuloBaseActivity.java
@@ -2,6 +2,7 @@ package com.stxnext.volontulo;
 
 import android.os.Bundle;
 import android.support.annotation.LayoutRes;
+import android.support.annotation.Nullable;
 import android.support.v7.app.ActionBar;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
@@ -11,6 +12,7 @@ import butterknife.ButterKnife;
 
 public abstract class VolontuloBaseActivity extends AppCompatActivity {
 
+    @Nullable
     @Bind(R.id.toolbar)
     protected Toolbar toolbar;
 
@@ -24,8 +26,10 @@ public abstract class VolontuloBaseActivity extends AppCompatActivity {
     }
 
     protected void init(String stringTitle) {
-        toolbar.setTitle(stringTitle);
-        setSupportActionBar(toolbar);
+        if (toolbar != null) {
+            toolbar.setTitle(stringTitle);
+            setSupportActionBar(toolbar);
+        }
 
         final ActionBar actionBar = getSupportActionBar();
         if (actionBar != null) {
@@ -37,6 +41,5 @@ public abstract class VolontuloBaseActivity extends AppCompatActivity {
     public void setContentView(@LayoutRes int layoutResID) {
         super.setContentView(layoutResID);
         ButterKnife.bind(this);
-        toolbar.setTitle(R.string.app_name);
     }
 }
