+++ Volontulo/app/src/main/java/com/stxnext/volontulo/ui/utils/BaseTextWatcher.java
@@ -1,4 +1,4 @@
-package com.stxnext.volontulo;
+package com.stxnext.volontulo.ui.utils;
 
 import android.text.Editable;
 import android.text.TextWatcher;
